+++ client/Application.jsx
@@ -20,7 +20,7 @@ import GameBoard from './GameBoard.jsx';
 import HowToPlay from './HowToPlay.jsx';
 import About from './About.jsx';
 import ForgotPassword from './pages/ForgotPassword.jsx';
-import ResetPassword from './ResetPassword.jsx';
+import ResetPassword from './pages/ResetPassword.jsx';
 import Profile from './Profile.jsx';
 import NewsAdmin from './NewsAdmin.jsx';
 import Unauthorised from './Unauthorised.jsx';
